# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Publication:
    properties:
      co_authors:
        description: Со-Авторы
        type: string
      impact:
        description: Влияние
        format: float
        type: number
      name:
        description: Название научной публикации
        type: string
      output_data:
        description: Выходные данные
        type: string
      publication_id:
        description: ID научной публикации
        format: uuid
        type: string
      rinc:
        description: Индекс РИНЦ
        type: boolean
      scopus:
        description: Индекс Scopus
        type: boolean
      status:
        description: Статус прогресса публикации
        enum:
        - to print
        - published
        - in progress
        type: string
      volume:
        description: Объем написанной работы
        type: integer
      wac:
        description: Индекс WaC
        type: boolean
      wos:
        description: Индекс WoS
        type: boolean
    type: object
  models.ResearchProject:
    properties:
      add_info:
        description: Дополнительная информация
        type: string
      end_at:
        description: Дата окончание
        format: date-time
        type: string
      grantee:
        description: Грантодатель
        type: string
      project_id:
        description: ID научного проекта
        format: uuid
        type: string
      project_name:
        description: Название проекта
        type: string
      start_at:
        description: Дата начала проекта
        format: date-time
        type: string
    type: object
...
  request_models.DeleteIDs:
    properties:
      ids:
        items:
          type: string
        type: array
      semester:
        type: integer
    type: object
...
  request_models.UpsertPublicationsRequest:
    properties:
      publications:
        description: |-
          Научные публикации, которые нужно добавить/изменить. Если запрос на добавление, то
          ID публикации (publication_id) не заполнять (делать null), в противном случае - иначе
        items:
          $ref: '#/definitions/models.Publication'
        type: array
      semester:
        type: integer
      work_id:
        description: ID совокупности всех научных работ за семестр
        format: uuid
        type: string
    type: object
  request_models.UpsertResearchProjectsRequest:
    properties:
      projects:
        description: |-
          Научные проекты, которые нужно добавить/изменить. Если запрос на добавление, то
          ID проекта (project_id) не заполнять (делать null), в противном случае - иначе
        items:
          $ref: '#/definitions/models.ResearchProject'
        type: array
      semester:
        type: integer
      work_id:
        description: ID совокупности всех научных работ за семестр
        format: uuid
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: Сервис Kasper
  version: 0.1.2
paths:
...
  /students/works/projects/{token}:
    post:
      consumes:
      - application/json
      description: Добавление или обновление научных проектов
      parameters:
      - description: Данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.UpsertResearchProjectsRequest'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Добавление или обновление научных проектов
      tags:
      - Student.ScientificWorks
    put:
      consumes:
      - application/json
      description: Удаление научных исследований по ID этих исследований
      parameters:
      - description: ID работ и семестр
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.DeleteIDs'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Удаление научных исследований
      tags:
      - Student.ScientificWorks
  /students/works/publications/{token}:
    post:
      consumes:
      - application/json
      description: Добавление или обновление научных публикаций
      parameters:
      - description: Данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.UpsertPublicationsRequest'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Добавление или обновление научных публикаций
      tags:
      - Student.ScientificWorks
    put:
      consumes:
      - application/json
      description: Удаление научных публикаций по ID публикаций
      parameters:
      - description: ID работ и семестр
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request_models.DeleteIDs'
      - description: Токен пользователя
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "204":
          description: Нет записей в БД
          schema:
            type: string
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "401":
          description: Токен протух
          schema:
            type: string
        "500":
          description: Ошибка на стороне сервера
          schema:
            type: string
      summary: Удаление научных публикаций
      tags:
      - Student.ScientificWorks
...
swagger: "2.0"